add pybind11 into cmake

+++ onnx/proto_utils.h
@@ -10,7 +10,7 @@ bool ParseProtoFromBytes(Proto* proto, const char* buffer, size_t length) {
   // Total bytes hard limit / warning limit are set to 1GB and 512MB
   // respectively.
   ::google::protobuf::io::CodedInputStream coded_stream(
-      new google::protobuf::io::ArrayInputStream(buffer, length));
+      new google::protobuf::io::ArrayInputStream(buffer, static_cast<int>(length)));
   coded_stream.SetTotalBytesLimit(1024LL << 20, 512LL << 20);
   return proto->ParseFromCodedStream(&coded_stream);
 }
